FamilyVoice/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
email-validator==1.0.5
eventlet==0.25.2
Flask==1.1.2
Flask-API==2.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.8.0
Flask-Login==0.5.0
Flask-Session==0.3.2
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Flask-Restful
greenlet==0.4.15
httplib2==0.9.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
Mako==1.0.7
MarkupSafe==1.1.1
numpy
PyJWT==1.7.1
pymacaroons==0.13.0
pymysql
pyRFC3339==1.0
python-dateutil==2.6.1
python-engineio==3.12.1
python-socketio==4.5.1
python-xlib==0.20
pytz==2018.3
redis==3.5.0
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
wadllib==1.3.2
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.1
